+++ kuStudy/kuStudyKit/model/PhotoLocation.swift
@@ -1,0 +1,14 @@
+//
+//  PhotoLocation.swift
+//  kuStudy
+//
+//  Created by BumMo Koo on 2016. 8. 10..
+//  Copyright © 2016년 gbmKSquare. All rights reserved.
+//
+
+import Foundation
+
+struct PhotoLocation {
+    let id: Int
+    let key: String
+}